Restructure testing folders
