+++ src/__test__/componentsTests/Header.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { MemoryRouter } from 'react-router-dom';
 
-import Header from '../components/Header';
+import Header from '../../components/Header';
 
 it('renders correctly Header component', () => {
   const header = renderer.create(
