+++ src/__test__/componentsTests/Missions.test.js
@@ -1,7 +1,7 @@
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import Missions from '../pages/Missions';
-import store from '../redux/configureStore';
+import Missions from '../../pages/Missions';
+import store from '../../redux/configureStore';
 
 describe('Missions page tests', () => {
   it('it renders', () => {
