+++ src/__test__/componentsTests/Profile.test.js
@@ -1,7 +1,7 @@
 import { Provider } from 'react-redux';
 import renderer from 'react-test-renderer';
-import Profile from '../pages/Profile';
-import store from '../redux/configureStore';
+import Profile from '../../pages/Profile';
+import store from '../../redux/configureStore';
 
 describe('Profile page', () => {
   it('it renders', () => {
