+++ src/__test__/componentsTests/Spinner.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Spinner from '../components/Spinner';
+import Spinner from '../../components/Spinner';
 
 it('renders correctly the Home component', () => {
   const spinner = renderer.create(<Spinner />).toJSON();
